treewide: move json-c compat shims into internal header file
Do not expose the json-c compat functions in ucode's public headers to
avoid clashes when building on systems with modern json-c.

Also remove some explicit json-c/json-c.h includes in places where it is
not needed.
